登录和注册功能

+++ src/main/java/com/inkyi/iblog/dao/InkUserMapper.java
@@ -5,13 +5,12 @@ import com.inkyi.iblog.entity.InkUserExample;
 import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
-public interface InkUserMapper extends BaseDao<InkUser, InkUserExample> {
-	
+public interface InkUserMapper extends BaseDao<InkUser, InkUserExample>{
     int countByExample(InkUserExample example);
 
     int deleteByExample(InkUserExample example);
 
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(InkUser record);
 
@@ -19,7 +18,7 @@ public interface InkUserMapper extends BaseDao<InkUser, InkUserExample> {
 
     List<InkUser> selectByExample(InkUserExample example);
 
-    InkUser selectByPrimaryKey(Integer id);
+    InkUser selectByPrimaryKey(Long id);
 
     int updateByExampleSelective(@Param("record") InkUser record, @Param("example") InkUserExample example);
